requirements.txt
fastapi
uvicorn[standard]
databases[postgresql]
sqlalchemy
psycopg2-binary
python-jose
pyjwt